Move color conversion helpers to a file
