+++ cardigan/stories/global/palette/palette.stories.tsx
@@ -3,6 +3,12 @@ import { themeValues } from '@weco/common/views/themes/config';
 import styled from 'styled-components';
 import { font } from '@weco/common/utils/classnames';
 import Divider from '@weco/common/views/components/Divider/Divider';
+import {
+  RGB,
+  HSL,
+  hexToRgb,
+  rgbToHsl,
+} from '@weco/common/utils/convert-colors';
 
 type PaletteColors = {
   [id: string]: Category;
@@ -18,18 +24,6 @@ type ColorObject = {
   hex: string;
   rgb: RGB;
   hsl: HSL;
-};
-
-type HSL = {
-  h: number;
-  s: number;
-  l: number;
-};
-
-type RGB = {
-  r: number;
-  g: number;
-  b: number;
 };
 
 const PaletteSection = styled.div`
@@ -78,60 +72,6 @@ const PaletteHex = styled.div.attrs({
 const PaletteCode = styled.code.attrs({
   className: font('lr', 6),
 })``;
-
-function hexToRgb(hex: string): RGB {
-  const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  return result
-    ? {
-        r: parseInt(result[1], 16),
-        g: parseInt(result[2], 16),
-        b: parseInt(result[3], 16),
-      }
-    : null;
-}
-
-// Code taken from https://css-tricks.com/converting-color-spaces-in-javascript/
-function rgbToHsl({ r, g, b }: RGB): HSL {
-  // Make r, g, and b fractions of 1
-  r /= 255;
-  g /= 255;
-  b /= 255;
-
-  // Find greatest and smallest channel values
-  const cmin = Math.min(r, g, b);
-  const cmax = Math.max(r, g, b);
-  const delta = cmax - cmin;
-  let h = 0;
-  let s = 0;
-  let l = 0;
-
-  // Calculate hue
-  // No difference
-  if (delta === 0) h = 0;
-  // Red is max
-  else if (cmax === r) h = ((g - b) / delta) % 6;
-  // Green is max
-  else if (cmax === g) h = (b - r) / delta + 2;
-  // Blue is max
-  else h = (r - g) / delta + 4;
-
-  h = Math.round(h * 60);
-
-  // Make negative hues positive behind 360°
-  if (h < 0) h += 360;
-
-  // Calculate lightness
-  l = (cmax + cmin) / 2;
-
-  // Calculate saturation
-  s = delta === 0 ? 0 : delta / (1 - Math.abs(2 * l - 1));
-
-  // Multiply l and s by 100
-  s = +(s * 100).toFixed(1);
-  l = +(l * 100).toFixed(1);
-
-  return { h, s, l };
-}
 
 let paletteColors: PaletteColors = {
   core: {
